+++ exspy/_signal_tools.py
